refactor: seprate CurriculumWeekCard logic
use .large for base view titles
