+++ Views/SettingsView.swift
@@ -61,6 +61,5 @@ struct SettingsView: View {
         }
         .scrollContentBackground(.hidden)
         .navigationTitle("Settings")
-        .navigationBarTitleDisplayMode(.inline)
     }
 }
